# AppWebCanvan/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Kanban API",
        "description": "API para gestionar tableros Kanban con columnas y tareas",
        "contact": {
            "email": "admin@kanban.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
        "/api/boards": {
            "get": {
                "tags": [
                    "Boards"
                ],
                "summary": "Obtener todos los tableros",
                "description": "Retorna una lista de todos los tableros con sus columnas y tareas",
                "operationId": "22c1b4dcc14110fc84dde21f6af5560b",
                "responses": {
                    "200": {
                        "description": "Lista de tableros obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "Desarrollo Web Laravel"
                                            },
                                            "descripcion": {
                                                "type": "string",
                                                "example": "Proyecto de desarrollo de aplicación web con Laravel"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "columns": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "board_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "titulo": {
                                                            "type": "string",
                                                            "example": "Backlog"
                                                        },
                                                        "color": {
                                                            "type": "string",
                                                            "example": "#FF6B6B"
                                                        },
                                                        "posicion": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "tasks": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "nombre": {
                                                                        "type": "string",
                                                                        "example": "Configurar autenticación"
                                                                    },
                                                                    "prioridad": {
                                                                        "type": "string",
                                                                        "example": "alta"
                                                                    },
                                                                    "avance": {
                                                                        "type": "integer",
                                                                        "example": 75
                                                                    },
                                                                    "status": {
                                                                        "type": "string",
                                                                        "example": "En curso"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Boards"
                ],
                "summary": "Crear un nuevo tablero",
                "description": "Crea un nuevo tablero Kanban",
                "operationId": "475fc0adeff918edaca9ab0b788b5e2b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Nuevo Tablero Kanban"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Descripción detallada del tablero de trabajo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tablero creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Nuevo Tablero Kanban"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Descripción detallada del tablero"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}